Add logic for Allocation and Chromosome assignment
Added ability to save vm and host pairing into an allocation. Each
allocation is stored in an arrayList inside Chromosome. Chromosome
represents one population; one population is the entire set of vms
allocated to vms randomly. Added methods to print all allocations to
view the entire set. Added the creation of a chromosome to the main
method.